+++ client/src/pages/Home/Home.js
@@ -1,9 +1,13 @@
 import React, { Component } from "react";
+// import  * from '../../utils';
+
 
 class All extends Component {
   render(){
     return(
       <h1>Welcome to my App!</h1>
+      
+      
     )
   }
 }
