--- controllers/firstController.js
@@ -1,19 +1,0 @@
-const db = require("../models");
-
-module.exports = {
-  findAll: function(req, res) {
-
-  },
-  findById: function(req, res) {
-
-  },
-  create: function(req, res) {
-
-  },
-  update: function(req, res) {
-
-  },
-  remove: function(req, res) {
-
-  }
-};
