+++ package.json
@@ -26,6 +26,8 @@
     "body-parser": "^1.17.2",
     "express": "^4.15.4",
     "mongoose": "^4.11.7",
+    "passport": "^0.4.0",
+    "passport-google-oauth": "^1.0.0",
     "react-document-title": "^2.0.3",
     "react-router-dom": "^4.3.1",
     "socket.io": "2.0",
