+++ server.js
@@ -6,6 +6,7 @@ const app = express();
 const PORT = process.env.PORT || 3001;
 var server = require('http').createServer(app)
 var io = require('socket.io')(server);
+var passport = require("passport");
 
 // Configure body parser for AJAX requests
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -14,6 +15,11 @@ app.use(bodyParser.json());
 app.use(express.static("client/build"));
 // Add routes, both API and view
 app.use(routes);
+
+
+app.use(passport.initialize()); // after line no.20 (express.static)
+
+require("./src/utils/passport");
 
 // Set up promises with mongoose
 mongoose.Promise = global.Promise;
